Started fillout out spar mode. For some reason, de-init crashes when unloading BGObject sprites

+++ main/modes/games/cGrove/Sparring/cg_Spar.h
@@ -22,4 +22,4 @@
 
 void cg_initSpar(cGrove_t* cg);
 void cg_deInitSpar(cGrove_t* cg);
-void cg_runSpar(cGrove_t* cg, int32_t elapsedUs);
+void cg_runSpar(cGrove_t* cg, int64_t elapsedUs);
